pyproject.toml
[project]
name = "nl2sql"
version = "0.1.0"
description = "自然語言轉 T-SQL 的 Web 應用程式"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Web UI
    "streamlit>=1.28.0",
    # Microsoft Agent Framework (預覽版)
    "agent-framework",
    # Azure 相關
    "azure-identity>=1.15.0",
    "openai>=1.12.0",
    # SQL Server 連線
    "pyodbc>=5.0.0",
    # 環境變數
    "python-dotenv>=1.0.0",
]

[project.scripts]
nl2sql = "app:main"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
]

